adds vscode .devcontainer development and code
- adds devcontainer.json to build and configure development container
- adds scripts to install and configure development tools
- moves sapb1 library code to src folder
