+++ .devcontainer/Dockerfile
@@ -1,0 +1,2 @@
+FROM mcr.microsoft.com/devcontainers/php
+EXPOSE 8080
