+++ .devcontainer/initializeCommand.shbat
@@ -1,0 +1,8 @@
+#!/usr/bin/env sh@ 2>/dev/null # 2>nul & echo off & goto BOF
+:echo "Linux env: executing .sh script"source initializeCommand.shexit
+:BOF
+@echo off
+@echo "Windows env: executing .bat script"
+@cmd /c initializeCommand.bat
+@exit /B %errorlevel%
+}
