+++ .env
@@ -1,0 +1,6 @@
+SAP_HOST=
+SAP_PORT=
+SAP_SSL=
+SAP_LOGIN=
+SAP_PASSWORD=
+SAP_COMPANY=
