--- SAPb1/Filters/Contains.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace SAPb1\Filters;
-
-class Contains extends Filter{
-    
-    private $field;
-    private $value;
-    
-    public function __construct($field, $value){
-        $this->field = $field;
-        $this->value = $value;
-    }
-
-    public function execute(){
-        return 'contains(' . $this->field . "," . $this->escape($this->value) . ")";
-    }
-}
