+++ initializeCommand.cmd
@@ -1,0 +1,5 @@
+#!/bin/bash
+# > nul 2>&1
+:; source .devcontainer/initializeCommand.sh;<<:
+@.devcontainer\initializeCommand.bat & ^
+:
