+++ src/SAPb1/Filters/index.php
@@ -1,0 +1,16 @@
+<?php 
+    require_once "Filter.php";
+    require_once "Between.php";
+    require_once "Contains.php";
+    require_once "EndsWith.php";
+    require_once "Equal.php";
+    require_once "InArray.php"    ;
+    require_once "LessThan.php";
+    require_once "LessThanEqual.php";
+    require_once "MoreThan.php";
+    require_once "MoreThanEqual.php";
+    require_once "NotEqual.php";
+    require_once "NotInArray.php";
+    require_once "Raw.php";
+    require_once "StartsWith.php";
+?>
